test(#877): extract the scene-readiness gate into a shared e2e util
Move the frame-settle wait out of the drag-pan benchmark into
src/wait-for-stable-frames.ts, matching the wait-for-honeypot-window
helper — it is scene-agnostic and the next live-canvas spec will reuse it.
